signup fix

+++ src/main/java/ddns/net/tracer/data/repository/LocationDataRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface LocationDataRepository extends CrudRepository<LocationData, Long> {
 
     LocationData findOneById(long id);
-    List<LocationData> findAllByTargetId(int id);
+    List<LocationData> findAllByTargetId(long id);
 
 
 }
