+++ src/main/java/ddns/net/tracer/data/service/TargetService.java
@@ -7,6 +7,6 @@ public interface TargetService {
 
     Target save(Target child);
     Target findOneByName(String name);
-    Target findOneById(int id);
+    Target findOneById(long id);
 
 }
