+++ src/main/java/ddns/net/tracer/data/service/TargetServiceImpl.java
@@ -35,7 +35,7 @@ public class TargetServiceImpl implements TargetService {
 
     @Transactional(readOnly = true)
     @Override
-    public Target findOneById(int id){
+    public Target findOneById(long id){
         return targetRepository.findOneById(id);
     }
